perf(iceberg): bound append manifest validation
